% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateMissingVariantCountStackBarChart.R
\name{generateMissingVariantCountStackBarChart}
\alias{generateMissingVariantCountStackBarChart}
\title{Create a missing variant count stack bar chart}
\usage{
generateMissingVariantCountStackBarChart(bcftools_tab_delimited_file_path)
}
\arguments{
\item{bcftools_tab_delimited_file_path}{A file path of bcftools tab delimited file.}
}
\value{
missing variant count stack bar chart.
}
\description{
The goal of generateMissingVariantCountStackBarChart is to create
a missing variant count stack bar chart for a bcftools processed
input tab delimited file.
}
\keyword{non_missing_variants}
