% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/generateMissingVariantCountStackBarChartParallel.R
\name{generateMissingVariantCountStackBarChartParallel}
\alias{generateMissingVariantCountStackBarChartParallel}
\title{Create a missing variant count stack bar chart}
\usage{
generateMissingVariantCountStackBarChartParallel(
  bcftools_tab_delimited_file_paths,
  cores = 1
)
}
\arguments{
\item{bcftools_tab_delimited_file_paths}{File paths of bcftools tab delimited files.}

\item{cores}{The number of cores used to process data.}
}
\value{
missing variant count stack bar chart.
}
\description{
The goal of generateMissingVariantCountStackBarChartParallel is to create
a missing variant count stack bar chart for a bcftools processed
input tab delimited file.
}
\keyword{non_missing_variants}
