% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/generateSampleMissingPercentageScatterPlotParallel.R
\name{generateSampleMissingPercentageScatterPlotParallel}
\alias{generateSampleMissingPercentageScatterPlotParallel}
\title{Create missing percentage scatter plot for samples in bcftools tab delimited file}
\usage{
generateSampleMissingPercentageScatterPlotParallel(
  bcftools_tab_delimited_file_paths,
  cores = 1
)
}
\arguments{
\item{bcftools_tab_delimited_file_paths}{File paths of bcftools tab delimited files.}

\item{cores}{The number of cores used to process data.}
}
\value{
sample missing percentage scatter plot.
}
\description{
The goal of generateSampleMissingPercentageScatterPlotParallel is to create
a missing percentage scatter plot for samples in bcftools tab delimited file.
}
\keyword{sample_missing_percentage_scatter_plot}
