% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateVariantCountHistogramParallel.R
\name{generateVariantCountHistogramParallel}
\alias{generateVariantCountHistogramParallel}
\title{Create a variant count histogram}
\usage{
generateVariantCountHistogramParallel(
  bcftools_tab_delimited_file_paths,
  cores = 1,
  gap = 5,
  binwidth = 1e+05
)
}
\arguments{
\item{bcftools_tab_delimited_file_paths}{File paths of bcftools tab delimited files.}

\item{cores}{The number of cores used to process data.}

\item{gap}{The gap of ticks in the x axis.}

\item{binwidth}{The width of each bin for the histogram.}
}
\value{
variant count histogram.
}
\description{
The goal of generateVariantCountHistogramParallel is to create
a variant count histogram for a bcftools processed input tab
delimited file.
}
\keyword{variant_count}
