% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualVariants-package.R
\docType{package}
\name{VisualVariants-package}
\alias{VisualVariants}
\alias{VisualVariants-package}
\title{VisualVariants: VisualVariants}
\description{
This package provides visualizations for variants.
}
\author{
\strong{Maintainer}: Yen On Chan \email{chanye@mail.missouri.edu} (\href{https://orcid.org/https://orcid.org/0000-0002-3345-9614}{ORCID})

}
\keyword{internal}
